+++ app/lib/LLRP/encode.js
