+++ app/lib/LLRP/messageConstants.js
