+++ app/lib/LLRP/parameterConstants.js
